updated unit tests

+++ src/Elsa.CustomActivities/Activities/Common/WeightedCheckboxModel.cs
@@ -20,5 +20,5 @@ namespace Elsa.CustomActivities.Activities.Common
 
     }
 
-    public record WeightedCheckboxRecord(string Identifier, string Answer, bool IsSingle, int Score, bool IsPrePopulated);
+    public record WeightedCheckboxRecord(string Identifier, string Answer, bool IsSingle, double Score, bool IsPrePopulated);
 }
