+++ src/Elsa.CustomActivities/Activities/Common/WeightedRadioModel.cs
@@ -5,5 +5,5 @@
         public ICollection<WeightedRadioRecord> Choices { get; set; } = new List<WeightedRadioRecord>();
     }
 
-    public record WeightedRadioRecord(string Identifier, string Answer, int Score, bool IsPrePopulated);
+    public record WeightedRadioRecord(string Identifier, string Answer, double Score, bool IsPrePopulated);
 }
